kumoko-net-cartographer/requirements.txt
dnspython
prettytable
